Completed implementation of pinned menu: fixed bug in decor.c where not all tiles would get deleted from the tbar; Implemented withdrawing of menu chain when one of them becomes pinned; Implemented proper tbar sizing for menus

+++ lib/decor.c
@@ -1104,7 +1104,7 @@ Bool
 delete_astbar_tile( ASTBarData *tbar, int idx )
 {
 
-    if( tbar != NULL && idx < tbar->tiles_num )
+    if( tbar != NULL && idx < (int)(tbar->tiles_num) )
     {
         register int i ;
         for( i = 0 ; i < tbar->tiles_num ; ++i )
